Fix README.md as the spec has changed
Didn't use redis in the end; sqlite would probably be fairly suitable
for this as it only needs to handle a single user; that'll be a coming
feature

Currently no persistence; as stated, that'll come later

Add a TODO since I'm currently marking notifications as complete but not
reading that anywhere - in fact I'm not attempting to do a real "sync"
or organisation of the notifications at all at the moment

Biggest issue with this is probably that every time the sync task ticks
around (every hour), we'll create a duplicate notification and ignore
the fact that we already have one. Top of the to do list.