Add a Routing config and basic recipient handling
This will allow the instance to identify which messages it should be
receiving and which it should be passing on

The instance is identified by a Recipient ID, and belongs to a number of
groups. It will display any notification directed at its ID, one of its
groups, or with no specified recipient.

If it knows an endpoint on which to contact a recipient or group, it
should pass on the notification to them, too. That'll be added later.