Rework REST a bit
No "propagate" boolean: always propagate if we think we know who the
destination should be.

Set the source of notifications to API if the "trigger" endpoint was
hit, though

Set up a "master" under routing in the config, indicating where we think
the master node of the cluster is. We'll assume it's us if that's blank

Add a "fetch" endpoint which will serve notifications for a particular
destination ID

Config loading FAILS if there's no master information set and we're not
set up to persist notifications: that makes us a master who isn't backed
by a database!

Add dev-master and dev-node configurations for testing

WIP: A couch view will be required for the fetchNotifications endpoint,
so that'll be in the next commit